Dow Jones Industrial Average
The Dow Jones Industrial Average (DJIA), Dow Jones, or simply the Dow (/ˈdaʊ/), is a stock market index of 30 prominent companies listed on stock exchanges in the United States.

The DJIA is one of the oldest and most commonly followed equity indexes. Many professionals consider it to be an inadequate representation of the overall U.S. stock market compared to a broader market index such as the S&P 500. The DJIA includes only 30 large companies. It is price-weighted, unlike stock indices, which use market capitalization. Furthermore, the DJIA does not use a weighted arithmetic mean.[4][5][6][7]

The value of the index can also be calculated as the sum of the stock prices of the companies included in the index, divided by a factor, which is currently (as of November 2021) approximately 0.152. The factor is changed whenever a constituent company undergoes a stock split so that the value of the index is unaffected by the stock split.

First calculated on May 26, 1896,[2] the index is the second-oldest among U.S. market indices, after the Dow Jones Transportation Average. It was created by Charles Dow, co-founder of both The Wall Street Journal and the Dow Jones & Company, and named after him and his business associate, statistician Edward Jones. The word "industrial" in the name of the index initially emphasized the heavy industry sector[dubious – discuss], but over time stocks from many other types of companies have been added to the DJIA.

The index is maintained by S&P Dow Jones Indices, an entity majority-owned by S&P Global. Its components are selected by a committee. The ten components with the largest dividend yields are commonly referred to as the Dogs of the Dow. As with all stock prices, the prices of the constituent stocks and consequently the value of the index itself are affected by the performance of the respective companies as well as macroeconomic factors.